Stop truncating operand pointers, and start closing the class at compile time
The expression evaluator pushed and popped operands through internals.intVal,
so every string operand lost its upper 32 bits twice -- on the way into the
stack and on the way out. A dvar name never reached Dvar_FindVar intact.

The larger finding is that CXX_FLAGS carried -w. The compiler has been
enumerating this entire class all along: 3654 pointer/int truncations across
460 of 666 translation units. Nearly every bug this port has chased for weeks
was a diagnostic we had silenced.

-w cannot be overridden by adding -Werror after it, so the ratchet removes it
per file instead. NX_SANITIZED_SOURCES starts with src/nx/, which needed three
fixes of its own before it compiled clean -- including a real LP64 truncation
in the audio handle and a rand() macro that rewrote std::rand() inside
libstdc++ templates.

The list is additive and the count prints at configure time, so a regression
is visible.

+++ src/sound/snd_driver_xaudio2.h
@@ -79,7 +79,10 @@ struct StreamVoice : IXAudio2VoiceCallback // sizeof=0x50
     // The buffer can now be reused or destroyed.
     STDMETHOD_(void, OnBufferEnd) (THIS_ void *pBufferContext)
     {
-        iSND_ReleaseStreamBuffer((unsigned int)pBufferContext >> 8, (unsigned __int8)pBufferContext);
+        // nx-port: the context is two small indices packed into the pointer, not
+        // an address -- go through uintptr_t so the cast is not a truncation.
+        iSND_ReleaseStreamBuffer((unsigned int)((uintptr_t)pBufferContext >> 8),
+                                 (unsigned __int8)(uintptr_t)pBufferContext);
     }
 
     // Called when this voice has just reached the end position of a loop.
